Add useOpen hook; streamline imports and state management
